Refactored EventsComponent out of the Admin mega component.
