+++ src/types/VenueEvent.ts
@@ -1,3 +1,6 @@
+import { WithId } from "utils/id";
+import { Venue } from "./Venue";
+
 export interface VenueEvent {
   name: string;
   start_utc_seconds: number;
@@ -9,3 +12,12 @@ export interface VenueEvent {
   host: string;
   room?: string;
 }
+
+export type VenueDetailsPartProps = {
+  venue: WithId<Venue>;
+  roomIndex?: number;
+  showCreateEventModal: boolean;
+  setShowCreateEventModal: Function;
+  editedEvent?: WithId<VenueEvent>;
+  setEditedEvent?: Function;
+};
